Added Gun Script and Bullet Script and 2D Zombie Asset Pack
The gun Script currently shoots the bullet right but that can be adjusted to shoot the bullet to where the gun is aiming. Just need to adjust one line of code. The bullet script simply detects collision with any object and returns itself to its start position (Behind the gun object or where the vector3 for the position is set. There are no enemies yet, but when we have enemies the damage to them will be done in the enemy script when colliding with the bullet using tags. Otherwise getting the enemy component value every time a bullet his him might be too heavy.