Added the Pests and their Animation
The player needs to be tagged as player for the pests to move to him. And in the script attached to the pest, when the pest is close enough to the player, it tries to modify the player's health to minus the amount of damage that the pest does. This will give an error because we don't have player health yet. But I tested with a dummy script and it works.