Added the attacking animations
Added the attacking animations and reversed the animation for the pests for now for demonstration. We can change this later when spawning the pests according to what we need. Also the animation for the pest when going over the background image it flashes sometimes as if the animation is going behind the image sometimes and then coming back to the front on the other frames.