Added AI Behavior and AI Gun and Adjusted Player Gun Script
The player gun script was shooting bullets but the bullet being shot war always horizontal and it was not facing the directing that it moving in. That it fixed now and the AI will aim and shoot the enemies closest to him if there are any, if not he will remain idle.